requirements.txt
requests>=2.31.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
cryptography>=41.0.0
tqdm>=4.66.0

# Forecasting pipeline
huggingface_hub>=0.23.0
sentence-transformers>=3.0.0
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.0.0
